+++ CHANGES.rst
@@ -36,6 +36,9 @@ New Features
 
 - Enable multiselect mode for selecting multiple table extensions in the Catalog loader. [#4051]
 
+- Table viewer tools to highlight selected, zoom-to-selected, and create subset from selected. [#3994]
+
+- Allow ingesting tables in plugins back into the app. [#4021]
 
 Cubeviz
 ^^^^^^^
@@ -123,6 +126,10 @@ Bug Fixes
 - Fixed astroquery and virtual observatory loaders not responding properly to link type changes. [#4038]
 
 - Avoid incorrectly showing "no viewer selected" warning when adding data from some plugins. [#4037]
+
+- Fix layer filtering logic for plot options to properly show/hide layers based on coordinate and link type. [#4046]
+
+- Avoid crashing Jdaviz if the PyVO vocabularies can't be downloaded. [#4059]
 
 Cubeviz
 ^^^^^^^